update
git-svn-id: https://svn.fe.up.pt/repos/lcom1415-t3g07@90 7d022d26-2eca-423b-9939-685a0713be1c

+++ lab5/sprite.c
@@ -40,7 +40,7 @@ int draw_sprite(Sprite *sp) {
 	int x = 1, y = 1;
 	char *map = sp->map;
 	while (x <= sp->width && y <= sp->height) {
-		draw_pixel(x + sp->x, y + sp->y, *map);
+		draw_pixel(x + (int)sp->x, y + (int)sp->y, *map);
 		if (x == sp->width) {
 			x = 0;
 			y++;
